+++ PeteBJJCompFeed/PeteBJJCompFeediOS/Feed UI/Composers/CompetitionsUIComposer.swift
@@ -40,26 +40,6 @@ private extension CompetitionsViewController {
     }
 }
 
-private final class WeakRefVirtualProxy<T: AnyObject> {
-    private weak var object: T?
-
-    init(_ object: T) {
-        self.object = object
-    }
-}
-
-extension WeakRefVirtualProxy: CompetitionsLoadingView where T: CompetitionsLoadingView {
-    func display(_ viewModel: CompetitionsLoadingViewModel) {
-        object?.display(viewModel)
-    }
-}
-
-extension WeakRefVirtualProxy: CompetitionsImageView where T: CompetitionsImageView, T.Image == UIImage {
-    func display(_ model: CompetitionsImageViewModel<UIImage>) {
-        object?.display(model)
-    }
-}
-
 private final class CompetitionsViewAdapter: CompetitionsView {
     private weak var controller: CompetitionsViewController?
     private let imageLoader: EventImageDataLoader
